pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "br-etl"
version = "0.1.0"
description = "ETL de dados de APIs governamentais brasileiras para investigação e análise"
readme = "README.md"
requires-python = ">=3.14"
keywords = ["etl", "dados-publicos", "transparencia", "brasil", "investigacao"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Utilities",
]
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[project.urls]
Homepage = "hhttps://github.com/brasilemdados/Olho-Cidadao"
Repository = "hhttps://github.com/brasilemdados/Olho-Cidadao"
Issues = "hhttps://github.com/brasilemdados/Olho-Cidadao/issues"
Documentation = "hhttps://github.com/brasilemdados/Olho-Cidadao/tree/main/docs"

[project.optional-dependencies]
dev = [
    "ruff>=0.15.5",
]

[project.scripts]
br-etl = "main:main"

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["cli*", "configuracao*", "extracao*", "infra*", "pipeline*", "utils*"]
namespaces = true

[tool.ruff]
target-version = "py314"
line-length = 100
